+++ CHANGELOG.md
@@ -30,6 +30,7 @@ For a full diff see [`88cf670...main`][88cf670...main].
 - Added `Tag` ([#292]), by [@localheinz]
 - Added `Reference` ([#293]), by [@localheinz]
 - Added `ReleaseList::first()` ([#297]), by [@localheinz]
+- Added `ReleaseList::last()` ([#298]), by [@localheinz]
 
 ### Changed
 
@@ -62,5 +63,6 @@ For a full diff see [`88cf670...main`][88cf670...main].
 [#295]: https://github.com/ergebnis/keep-a-changelog/pull/295
 [#296]: https://github.com/ergebnis/keep-a-changelog/pull/296
 [#297]: https://github.com/ergebnis/keep-a-changelog/pull/297
+[#298]: https://github.com/ergebnis/keep-a-changelog/pull/298
 
 [@localheinz]: https://github.com/localheinz